factor LimitTrait and LimitOffsetTrait into LimitInterface and LimitOffsetInterface; clean up
