+++ src/Common/LimitInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Aura\Sql_Query\Common;
+
+interface LimitInterface
+{
+    /**
+     *
+     * Sets a limit count on the query.
+     *
+     * @param int $limit The number of rows to select.
+     *
+     * @return $this
+     *
+     */
+    public function limit($limit);
+}
