+++ src/Common/Select.php
@@ -12,7 +12,6 @@ namespace Aura\Sql_Query\Common;
 
 use Aura\Sql_Query\AbstractQuery;
 use Aura\Sql_Query\Exception;
-use Aura\Sql_Query\Traits;
 
 /**
  *
@@ -23,8 +22,6 @@ use Aura\Sql_Query\Traits;
  */
 class Select extends AbstractQuery implements SelectInterface
 {
-    use Traits\LimitOffsetTrait;
-
     /**
      *
      * An array of union SELECT statements.
@@ -637,6 +634,36 @@ class Select extends AbstractQuery implements SelectInterface
 
         $this->addWhere($cond, 'OR', $bind);
 
+        return $this;
+    }
+
+    /**
+     *
+     * Sets a limit count on the query.
+     *
+     * @param int $limit The number of rows to select.
+     *
+     * @return $this
+     *
+     */
+    public function limit($limit)
+    {
+        $this->limit = (int) $limit;
+        return $this;
+    }
+
+    /**
+     *
+     * Sets a limit offset on the query.
+     *
+     * @param int $offset Start returning after this many rows.
+     *
+     * @return $this
+     *
+     */
+    public function offset($offset)
+    {
+        $this->offset = (int) $offset;
         return $this;
     }
